# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: National Public Asset Management System (NPAMS) API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              required: [refresh_token]
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
  /v1/auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /v1/auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        "200":
          description: Reset link sent if email exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /v1/categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List asset categories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AssetCategory"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create asset category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category_name:
                  type: string
                description:
                  type: string
              required: [category_name]
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /v1/categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category_name:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /v1/dashboard/provincial:
    get:
      operationId: getProvincialDashboard
      tags: [dashboard]
      summary: Provincial dashboard summary
      security:
        - bearerAuth: []
      parameters:
        - name: province_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Provincial dashboard data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/ProvincialDashboard"
  /v1/dashboard/national:
    get:
      operationId: getNationalDashboard
      tags: [dashboard]
      summary: National dashboard summary (national roles only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: National dashboard data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/NationalDashboard"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    UserScope:
      type: object
      properties:
        province_id:
          type: string
          nullable: true
        district_id:
          type: string
          nullable: true
        facility_id:
          type: string
          nullable: true
    LoginUser:
      type: object
      properties:
        id:
          type: string
        full_name:
          type: string
        email:
          type: string
        role:
          type: string
        scope_level:
          type: string
        scope:
          $ref: "#/components/schemas/UserScope"
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
            expires_in:
              type: integer
            user:
              $ref: "#/components/schemas/LoginUser"
    RefreshResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            access_token:
              type: string
            expires_in:
              type: integer
...
    AssetCategory:
      type: object
      properties:
        id:
          type: string
        categoryName:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    CategoryCount:
      type: object
      properties:
        category_name:
          type: string
          nullable: true
        count:
          type: integer
    ConditionCount:
      type: object
      properties:
        condition:
          type: string
          nullable: true
        count:
          type: integer
    RecentAsset:
      type: object
      properties:
        id:
          type: string
        assetTag:
          type: string
        assetName:
          type: string
        status:
          type: string
        condition:
          type: string
        createdAt:
          type: string
          format: date-time
        categoryName:
          type: string
          nullable: true
        facilityName:
          type: string
          nullable: true
    ProvinceInfo:
      type: object
      nullable: true
      properties:
        provinceName:
          type: string
        flagUrl:
          type: string
          nullable: true
        themeAccentColor:
          type: string
          nullable: true
    ProvincialDashboard:
      type: object
      properties:
        province:
          $ref: "#/components/schemas/ProvinceInfo"
        total_assets:
          type: integer
        active_assets:
          type: integer
        missing_assets:
          type: integer
        total_value:
          type: string
        assets_by_category:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
        assets_by_condition:
          type: array
          items:
            $ref: "#/components/schemas/ConditionCount"
        recent_assets:
          type: array
          items:
            $ref: "#/components/schemas/RecentAsset"
    ProvinceAssetSummary:
      type: object
      properties:
        province_id:
          type: string
        province_name:
          type: string
        province_code:
          type: string
        flag_url:
          type: string
          nullable: true
        theme_accent_color:
          type: string
          nullable: true
        total_assets:
          type: integer
        total_value:
          type: string
        missing_assets:
          type: integer
    NationalDashboard:
      type: object
      properties:
        total_assets:
          type: integer
        total_value:
          type: string
        provinces_count:
          type: integer
        assets_by_province:
          type: array
          items:
            $ref: "#/components/schemas/ProvinceAssetSummary"
        top_categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
...