# lib/api-zod/src/generated/types/conditionCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * National Public Asset Management System (NPAMS) API
 * OpenAPI spec version: 1.0.0
 */

export interface ConditionCount {
  condition?: string | null;
  count?: number;
}