# lib/api-zod/src/generated/types/recentAsset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * National Public Asset Management System (NPAMS) API
 * OpenAPI spec version: 1.0.0
 */

export interface RecentAsset {
  id?: string;
  assetTag?: string;
  assetName?: string;
  status?: string;
  condition?: string;
  createdAt?: Date;
  categoryName?: string | null;
  facilityName?: string | null;
}